Excel Dashboard/requirements.txt
Flask
pandas
requests
python-dateutil
openpyxl
serverless-wsgi
